Add the Track function to the log to track events.
